==> code/05_major_waiting_times.R <==
library(tidyverse)
library(ggplot2)
library(patchwork)

player_penalties <- read.csv("data/plays_players.csv")
penalties <- read.csv("data/game_penalties.csv")
players <- read.csv("data/player_info.csv")
plays_players1 <- readRDS("data/plays_players1.rds")
games <- read.csv("data/games.csv")

filtered_penalties <- readRDS("intermediate_data/filtered_penalties.rds")
games_ordered <- readRDS("data/games_ordered.rds")
complete_penalty_info <- readRDS("intermediate_data/complete_penalty_info.rds")
all_player_plays <- readRDS("intermediate_data/all_player_plays.rds")
careers <- readRDS("intermediate_data/careers.rds")
career_data <- readRDS("intermediate_data/career_data.rds")
mean_share_table <- readRDS("intermediate_data/mean_share_table.rds")
player_games <- readRDS("intermediate_data/player_games.rds")
tab_games_by_players <- readRDS("intermediate_data/tab_games_by_players.rds")
tab_seasons_by_players <- readRDS("intermediate_data/tab_seasons_by_players.rds")
tab_penalties_by_players <- readRDS("intermediate_data/tab_penalties_by_players.rds")
normalized_player_data <- readRDS("intermediate_data/normalized_player_data.rds")
averages <- readRDS("intermediate_data/averages.rds")
player_game_count <- readRDS("intermediate_data/player_game_count.rds")

source("code/function_waiting_times.R")
source("code/function_most_common_num_pens.R")
source("code/function_fixednumber_penalties_season.R")
source("code/function_season_plays.R")
source("code/function_game_count_season.R")
source("code/function_player_games_season.R")

# # all of the steps to make the plot for one waiting time. This is what is used for all of the different penalty amounts, the specific year and index data frame just has to be brought in
# 
# # calling the list of data frames created by the waiting_times function for the season and then going data frame by data frame in the list of waiting times
# majorwaitingtimesseason <- waiting_times(year, "Major") (put specific season in)
# majorwaitingtimesseason_index <- as.data.frame(majorwaitingtimes20192020[index]) (index in the waiting_times list- 1 is the most frequent number of repeat major penalties, 2 is second most, 3 is third most)
# 
# # the highest game count for a player in the given data frame, used to determine the axis of the plot and the largest possible waiting time to determine how many rows are needed in the majorwaitingtimesdf
# most_games <- max(majorwaitingtimesseason_index$game_count)
# 
# # dataframe counting the observed amount of each observed waiting time for the plot
# counts_obs <- majorwaitingtimesseason_index %>%
#   group_by(game_diff) %>%
#   summarize(obs_count = n())
# 
# # main dataframe that will have all possible waiting times and counts of the observed and random iterations. this can be the same for all seasons when you do it
# majorwaitingtimes <- data.frame(game_diff = 0:most_games)
# majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
# # if there is a game difference in the range that is not observed, it will come up as NA. since we want to look at all waiting times, this command changes the NA to 0
# for (diff in 1:nrow(majorwaitingtimes)) {
#   if (is.na(majorwaitingtimes$obs_count[diff])) {
#     majorwaitingtimes$obs_count[diff] <- 0
#   } 
# }
# 
# 
# #loop picking random games out of the games each player played each season to have penalties in
# # can set the amount of loops higher, will have to change the range of values below to the new upper bound + 2. ie in this case we are looking at i in 1:100, so the majorwaitingtimes data frame will have 102 columns (the 100 random plus the observed and the player ids)
# # list that will hold the results of all 100 iterations
# randommajorwaitingtimes_list <- list()
# for (i in 1:100) {
#   # list that will hold the results of each player in each iteration
#   randommajorwaitingtimes1_list <- list()
#   for (player in majorwaitingtimesseason_index$player_id) {
#     num_games <- majorwaitingtimesseason_index$game_count[majorwaitingtimesseason_index$player_id == player] (each player's game count)
#     num_penalties <- sum(majorwaitingtimesseason_index$player_id == player) + 1 (# of pens each player had in the season. will be the same for each player from previous filtering. + 1 b/c majorwaitingtimesseason_index amount of waiting tims is the amount of penalties - 1)
#     randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
#     randomgames_ordered <- sort(randomgames)
#     player_diffs <- filter(majorwaitingtimesseason_index, player_id == player) (create a data frame for each specfic player and then will be combined later)
#     
#     randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
#     for (game in 1:(num_penalties - 1)){ ( -1 since the last penalty does not have a waiting time after it)
#       penalty_game <- randomgames_ordered[game] (going through each random game and calculating waiting times, then adding to df and storing in the list for each player)
#       randommajorwaitingtimes$penalty_game[game] <- penalty_game
#       
#       next_penalty_game <- randomgames_ordered[game + 1]
#       randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
#       
#       game_diff <- abs(next_penalty_game - penalty_game)
#       randommajorwaitingtimes$game_diff[game] <- game_diff
#       
#       randommajorwaitingtimes$game_count[game] <- num_games
#     }
#     randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
#   }
#   
#   
#   randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
#   randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
#   
#   
#   game_counts <- randommajorwaitingtimes_list[[i]] %>% (game_diff counts for each iteration)
#     group_by(game_diff) %>%
#     summarize(count = n())
#   
#   counts <- data.frame(game_diff = 0:most_games)
#   counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
#   for (diff in 1:nrow(majorwaitingtimes)) {
#     if (is.na(counts$count[diff])) {
#       counts$count[diff] <- 0
#     }
#   }
#   colname <- paste0("count_simulation_", i) (adding the counts for each iteration to the main df)
#   majorwaitingtimes[[colname]] <- counts$count
# }
# 
# # loop for getting the upper and lower bounds for the plot
# quantile_list <- list()
# 
# # pulling the values from each row (which are the counts for each waiting time from each iteration)
# for (i in 1:nrow(majorwaitingtimes)) {
#   values <- as.numeric(majorwaitingtimes[i, 3:102])
#   quantiles <- quantile(values, probs = c(0.025, 0.975))
#   quantile_list[[i]] <- quantiles
# }
# 
# quantiles_df <- do.call(rbind, quantile_list) (combining all of the individual quantiles so everything is in one df)
# # making a data frame with all of the simulation values and quantiles
# obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
# majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)
# 
# majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
#   rename(lower_bound = "2.5%",
#          upper_bound = "97.5%")
# 
# p <- ggplot(data = majorwaitingtimes_quantiles) + 
#   geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
#   geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
#   geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
#   geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
#   geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
#   geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
#   scale_y_continuous(breaks = (0:100)) +
#   scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
#   labs(title = paste0("Waiting Times of Players with ", num_penalties, "Major Penalties in the year-year Season"), x = "Game difference", y = "Count of players")+
#   theme_minimal() +
#   theme(panel.grid.minor = element_blank())
# 
# p
# 
# ggsave("results/majorwaitingtimesseason_index.png",p,bg="white")
# 
# # stops here - this does everything to create the plot for the waiting times between penalties for players with the most common penalty amount

#2019-2020
#waiting times for the top 3 most frequent amount of penalties for players to have
majorwaitingtimes20192020 <- waiting_times(2019, "Major")
majorwaitingtimes20192020_1 <- as.data.frame(majorwaitingtimes20192020[1])

most_games <- max(majorwaitingtimes20192020_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20192020_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20192020_1$player_id) {
    num_games <- majorwaitingtimes20192020_1$game_count[majorwaitingtimes20192020_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20192020_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20192020_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2019-2020 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20192020_1.png",p,bg="white")

#2019-2020
#waiting times for players with the second most common penalty amount

majorwaitingtimes20192020_2 <- as.data.frame(majorwaitingtimes20192020[2])

most_games <- max(majorwaitingtimes20192020_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20192020_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20192020_2$player_id) {
    num_games <- majorwaitingtimes20192020_2$game_count[majorwaitingtimes20192020_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20192020_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20192020_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2019-2020 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20192020_2.png",p,bg="white")

#2019-2020
#waiting times for players with the third most common penalty amount

majorwaitingtimes20192020_3 <- as.data.frame(majorwaitingtimes20192020[3])

most_games <- max(majorwaitingtimes20192020_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20192020_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20192020_3$player_id) {
    num_games <- majorwaitingtimes20192020_3$game_count[majorwaitingtimes20192020_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20192020_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20192020_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2019-2020 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20192020_3.png",p,bg="white")


#2018-2019
#waiting times for players with the most common penalty amount

majorwaitingtimes20182019 <- waiting_times(2018, "Major")
majorwaitingtimes20182019_1 <- as.data.frame(majorwaitingtimes20182019[1])

most_games <- max(majorwaitingtimes20182019_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20182019_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20182019_1$player_id) {
    num_games <- majorwaitingtimes20182019_1$game_count[majorwaitingtimes20182019_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20182019_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20182019_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2018-2019 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20182019_1.png",p,bg="white")

#2018-2019
#waiting times for players with the second most common penalty amount

majorwaitingtimes20182019_2 <- as.data.frame(majorwaitingtimes20182019[2])

most_games <- max(majorwaitingtimes20182019_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20182019_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20182019_2$player_id) {
    num_games <- majorwaitingtimes20182019_2$game_count[majorwaitingtimes20182019_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20182019_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20182019_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2018-2019 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20182019_2.png",p,bg="white")

#2018-2019
#waiting times for players with the third most common penalty amount

majorwaitingtimes20182019_3 <- as.data.frame(majorwaitingtimes20182019[3])

most_games <- max(majorwaitingtimes20182019_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20182019_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20182019_3$player_id) {
    num_games <- majorwaitingtimes20182019_3$game_count[majorwaitingtimes20182019_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20182019_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20182019_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2018-2019 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20182019_3.png",p,bg="white")



#2017-2018
#waiting times for players with the most common penalty amount

majorwaitingtimes20172018 <- waiting_times(2017, "Major")
majorwaitingtimes20172018_1 <- as.data.frame(majorwaitingtimes20172018[1])

most_games <- max(majorwaitingtimes20172018_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20172018_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20172018_1$player_id) {
    num_games <- majorwaitingtimes20172018_1$game_count[majorwaitingtimes20172018_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20172018_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20172018_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2017-2018 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20172018_1.png",p,bg="white")

#2017-2018
#waiting times for players with the second most common penalty amount

majorwaitingtimes20172018_2 <- as.data.frame(majorwaitingtimes20172018[2])

most_games <- max(majorwaitingtimes20172018_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20172018_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20172018_2$player_id) {
    num_games <- majorwaitingtimes20172018_2$game_count[majorwaitingtimes20172018_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20172018_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20172018_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2017-2018 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20172018_2.png",p,bg="white")

#2017-2018
#waiting times for players with the third most common penalty amount

majorwaitingtimes20172018_3 <- as.data.frame(majorwaitingtimes20172018[3])

most_games <- max(majorwaitingtimes20172018_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20172018_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20172018_3$player_id) {
    num_games <- majorwaitingtimes20172018_3$game_count[majorwaitingtimes20172018_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20172018_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20172018_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2017-2018 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20172018_3.png",p,bg="white")


#2016-2017
#waiting times for players with the most common penalty amount

majorwaitingtimes20162017 <- waiting_times(2016, "Major")
majorwaitingtimes20162017_1 <- as.data.frame(majorwaitingtimes20162017[1])

most_games <- max(majorwaitingtimes20162017_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20162017_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20162017_1$player_id) {
    num_games <- majorwaitingtimes20162017_1$game_count[majorwaitingtimes20162017_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20162017_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20162017_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2016-2017 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20162017_1.png",p,bg="white")

#2016-2017
#waiting times for players with the second most common penalty amount

majorwaitingtimes20162017_2 <- as.data.frame(majorwaitingtimes20162017[2])

most_games <- max(majorwaitingtimes20162017_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20162017_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20162017_2$player_id) {
    num_games <- majorwaitingtimes20162017_2$game_count[majorwaitingtimes20162017_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20162017_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20162017_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2016-2017 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20162017_2.png",p,bg="white")

#2016-2017
#waiting times for players with the third most common penalty amount

majorwaitingtimes20162017_3 <- as.data.frame(majorwaitingtimes20162017[3])

most_games <- max(majorwaitingtimes20162017_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20162017_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20162017_3$player_id) {
    num_games <- majorwaitingtimes20162017_3$game_count[majorwaitingtimes20162017_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20162017_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20162017_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2016-2017 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20162017_3.png",p,bg="white")


#2015-2016
#waiting times for players with the most common penalty amount

majorwaitingtimes20152016 <- waiting_times(2015, "Major")
majorwaitingtimes20152016_1 <- as.data.frame(majorwaitingtimes20152016[1])

most_games <- max(majorwaitingtimes20152016_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20152016_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20152016_1$player_id) {
    num_games <- majorwaitingtimes20152016_1$game_count[majorwaitingtimes20152016_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20152016_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20152016_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2015-2016 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20152016_1.png",p,bg="white")

#2015-2016
#waiting times for players with the second most common penalty amount

majorwaitingtimes20152016_2 <- as.data.frame(majorwaitingtimes20152016[2])

most_games <- max(majorwaitingtimes20152016_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20152016_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20152016_2$player_id) {
    num_games <- majorwaitingtimes20152016_2$game_count[majorwaitingtimes20152016_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20152016_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20152016_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2015-2016 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20152016_2.png",p,bg="white")

#2015-2016
#waiting times for players with the third most common penalty amount

majorwaitingtimes20152016_3 <- as.data.frame(majorwaitingtimes20152016[3])

most_games <- max(majorwaitingtimes20152016_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20152016_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20152016_3$player_id) {
    num_games <- majorwaitingtimes20152016_3$game_count[majorwaitingtimes20152016_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20152016_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20152016_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2015-2016 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20152016_3.png",p,bg="white")


#2014-2015
# waiting times for players with the most common penalty amount

majorwaitingtimes20142015 <- waiting_times(2014, "Major")
majorwaitingtimes20142015_1 <- as.data.frame(majorwaitingtimes20142015[1])

most_games <- max(majorwaitingtimes20142015_1$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20142015_1 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20142015_1$player_id) {
    num_games <- majorwaitingtimes20142015_1$game_count[majorwaitingtimes20142015_1$player_id == player]
    num_penalties <- sum(majorwaitingtimes20142015_1$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20142015_1, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2014-2015 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20142015_1.png",p,bg="white")

#2014-2015
#waiting times for players with the second most common penalty amount

majorwaitingtimes20142015_2 <- as.data.frame(majorwaitingtimes20142015[2])

most_games <- max(majorwaitingtimes20142015_2$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20142015_2 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20142015_2$player_id) {
    num_games <- majorwaitingtimes20142015_2$game_count[majorwaitingtimes20142015_2$player_id == player]
    num_penalties <- sum(majorwaitingtimes20142015_2$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20142015_2, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2014-2015 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20142015_2.png",p,bg="white")

#2014-2015
#waiting times for players with the third most common penalty amount

majorwaitingtimes20142015_3 <- as.data.frame(majorwaitingtimes20142015[3])

most_games <- max(majorwaitingtimes20142015_3$game_count)

# dataframe counting the observed counts of each observed waiting time
counts_obs <- majorwaitingtimes20142015_3 %>%
  group_by(game_diff) %>%
  summarize(obs_count = n())

# main dataframe that will have all possible waiting times and counts
majorwaitingtimes <- data.frame(game_diff = 0:most_games)
majorwaitingtimes <- merge(majorwaitingtimes, counts_obs, by = "game_diff", all.x = TRUE)
for (diff in 1:nrow(majorwaitingtimes)) {
  if (is.na(majorwaitingtimes$obs_count[diff])) {
    majorwaitingtimes$obs_count[diff] <- 0
  } 
}


#loop picking random games out of the games they played each season to have a penalty in 
randommajorwaitingtimes_list <- list()
for (i in 1:100) {
  randommajorwaitingtimes1_list <- list()
  for (player in majorwaitingtimes20142015_3$player_id) {
    num_games <- majorwaitingtimes20142015_3$game_count[majorwaitingtimes20142015_3$player_id == player]
    num_penalties <- sum(majorwaitingtimes20142015_3$player_id == player) + 1
    randomgames <- sample(1:num_games, num_penalties, replace = TRUE)
    randomgames_ordered <- sort(randomgames)
    player_diffs <- filter(majorwaitingtimes20142015_3, player_id == player)
    
    randommajorwaitingtimes <- data.frame(player_id = player_diffs$player_id, penalty_game = NA, next_penalty_game = NA, game_diff = NA, game_count = NA)
    for (game in 1:(num_penalties - 1)){
      penalty_game <- randomgames_ordered[game]
      randommajorwaitingtimes$penalty_game[game] <- penalty_game
      
      next_penalty_game <- randomgames_ordered[game + 1]
      randommajorwaitingtimes$next_penalty_game[game] <- next_penalty_game
      
      game_diff <- abs(next_penalty_game - penalty_game)
      randommajorwaitingtimes$game_diff[game] <- game_diff
      
      randommajorwaitingtimes$game_count <- num_games
    }
    randommajorwaitingtimes1_list[[player]] <- randommajorwaitingtimes
  }
  
  randommajorwaitingtimes1 <- do.call(rbind, randommajorwaitingtimes1_list)
  randommajorwaitingtimes_list[[i]] <- as.data.frame(randommajorwaitingtimes1)
  
  
  game_counts <- randommajorwaitingtimes_list[[i]] %>%
    group_by(game_diff) %>%
    summarize(count = n())
  
  counts <- data.frame(game_diff = 0:most_games)
  counts <- merge(counts, game_counts, by = "game_diff", all.x = TRUE)
  for (diff in 1:nrow(majorwaitingtimes)) {
    if (is.na(counts$count[diff])) {
      counts$count[diff] <- 0
    }
  }
  colname <- paste0("count_simulation_", i)
  majorwaitingtimes[[colname]] <- counts$count
}

# loop for getting the upper and lower bounds
quantile_list <- list()

# pulling the values from each row (which are the counts for each waiting time from each iteration)
for (i in 1:nrow(majorwaitingtimes)) {
  values <- as.numeric(majorwaitingtimes[i, 3:102])
  quantiles <- quantile(values, probs = c(0.025, 0.975))
  quantile_list[[i]] <- quantiles
}

quantiles_df <- do.call(rbind, quantile_list)
# making a data frame with all of the simulation values and quantiles
obsmajorwaitingtimes <- majorwaitingtimes[, c("game_diff", "obs_count")]
majorwaitingtimes_quantiles <- cbind(obsmajorwaitingtimes, quantiles_df)

majorwaitingtimes_quantiles <- majorwaitingtimes_quantiles %>%
  rename(lower_bound = "2.5%",
         upper_bound = "97.5%")

p <- ggplot(data = majorwaitingtimes_quantiles) + 
  geom_point(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_line(mapping = aes(x = game_diff, y = obs_count, group=1, color = "Observed")) +
  geom_point(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_line(mapping = aes(x = game_diff, y = lower_bound, group=1, color = "Lower Bound")) +
  geom_point(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  geom_line(mapping = aes(x = game_diff, y = upper_bound, group=1, color = "Upper Bound")) +
  scale_y_continuous(breaks = (0:35)) +
  scale_x_continuous(breaks = seq(0, most_games, by = 2)) +
  labs(title = paste0("Waiting Times of Players with ", num_penalties, " Major Penalties in the 2014-2015 Season"), x = "Game difference", y = "Count of players")+
  theme_minimal() +
  theme(panel.grid.minor = element_blank())

p

ggsave("results/majorwaitingtimes20142015_3.png",p,bg="white")
